Add files via upload
The dataset used for the entirety of the project. Contains over 7,000 rows of data about the salaries from ai-jobs. Ai-jobs collects salary information anonymously from professionals all over the world in the AI/ML and Big Data space and makes it publicly available for anyone to use, share and play around with.